Move action types to /ts/store/actions
