+++ ts/boot/configureStoreAndPersistor.ts
@@ -12,11 +12,11 @@ import storage from "redux-persist/lib/storage";
 import createSagaMiddleware from "redux-saga";
 import thunk from "redux-thunk";
 
-import { Action, Store, StoreEnhancer } from "../actions/types";
 import { analytics } from "../middlewares";
 import rootReducer from "../reducers";
 import { GlobalState } from "../reducers/types";
 import rootSaga from "../sagas";
+import { Action, Store, StoreEnhancer } from "../store/actions/types";
 import { NAVIGATION_MIDDLEWARE_LISTENERS_KEY } from "../utils/constants";
 
 const isDebuggingInChrome = __DEV__ && !!window.navigator.userAgent;
