+++ ts/reducers/types.ts
@@ -1,7 +1,7 @@
 import { NavigationState } from "react-navigation";
 import { FormStateMap } from "redux-form";
 
-import { Action } from "../actions/types";
+import { Action } from "../store/actions/types";
 import { EntitiesState } from "../store/reducers/entities";
 import { ErrorState } from "../store/reducers/error";
 import { LoadingState } from "../store/reducers/loading";
