+++ ts/screens/main/MessagesScreen.tsx
@@ -8,11 +8,11 @@ import {
 } from "react-navigation";
 import { connect } from "react-redux";
 
-import { ReduxProps } from "../../actions/types";
 import I18n from "../../i18n";
 import { GlobalState } from "../../reducers/types";
 import { FetchRequestActions } from "../../store/actions/constants";
 import { loadMessages } from "../../store/actions/messages";
+import { ReduxProps } from "../../store/actions/types";
 import { orderedMessagesSelector } from "../../store/reducers/entities/messages";
 import { createLoadingSelector } from "../../store/reducers/loading";
 import variables from "../../theme/variables";
