+++ ts/screens/onboarding/PinScreen.tsx
@@ -15,13 +15,13 @@ import { NavigationScreenProp, NavigationState } from "react-navigation";
 import { connect } from "react-redux";
 
 import { Option } from "fp-ts/lib/Option";
-import { ReduxProps } from "../../actions/types";
 import Pinpad from "../../components/Pinpad";
 import AppHeader from "../../components/ui/AppHeader";
 import TextWithIcon from "../../components/ui/TextWithIcon";
 import I18n from "../../i18n";
 import { GlobalState } from "../../reducers/types";
 import { createPin } from "../../store/actions/onboarding";
+import { ReduxProps } from "../../store/actions/types";
 import { createErrorSelector } from "../../store/reducers/error";
 
 type ReduxMappedProps = {
