+++ ts/screens/onboarding/TosScreen.tsx
@@ -13,10 +13,10 @@ import {
 import * as React from "react";
 import { NavigationScreenProp, NavigationState } from "react-navigation";
 import { connect } from "react-redux";
-import { ReduxProps } from "../../actions/types";
 import AppHeader from "../../components/ui/AppHeader";
 import I18n from "../../i18n";
 import { acceptTos } from "../../store/actions/onboarding";
+import { ReduxProps } from "../../store/actions/types";
 type ReduxMappedProps = {};
 type OwnProps = {
   navigation: NavigationScreenProp<NavigationState>;
