+++ ts/store/actions/types.ts
@@ -1,5 +1,5 @@
 /**
- * Defines Flow types for the available actions and store related stuff.
+ * Defines types for the available actions and store related stuff.
  */
 
 import { NavigationAction } from "react-navigation";
@@ -10,16 +10,16 @@ import {
   StoreEnhancer as ReduxStoreEnhancer
 } from "redux";
 
-import { GlobalState } from "../reducers/types";
-import { ApplicationActions } from "../store/actions/application";
-import { APP_STATE_CHANGE_ACTION } from "../store/actions/constants";
-import { ErrorActions } from "../store/actions/error";
-import { MessagesActions } from "../store/actions/messages";
-import { NotificationsActions } from "../store/actions/notifications";
-import { OnboardingActions } from "../store/actions/onboarding";
-import { ProfileActions } from "../store/actions/profile";
-import { ServicesActions } from "../store/actions/services";
-import { SessionActions } from "../store/actions/session";
+import { GlobalState } from "../../reducers/types";
+import { ApplicationActions } from "./application";
+import { APP_STATE_CHANGE_ACTION } from "./constants";
+import { ErrorActions } from "./error";
+import { MessagesActions } from "./messages";
+import { NotificationsActions } from "./notifications";
+import { OnboardingActions } from "./onboarding";
+import { ProfileActions } from "./profile";
+import { ServicesActions } from "./services";
+import { SessionActions } from "./session";
 
 export type ApplicationState = "background" | "inactive" | "active";
