+++ ts/store/reducers/entities/index.ts
@@ -3,7 +3,7 @@
  */
 import { combineReducers } from "redux";
 
-import { Action } from "../../../actions/types";
+import { Action } from "../../actions/types";
 import messagesReducer, { MessagesState } from "./messages";
 import servicesReducer, { ServicesState } from "./services";
