+++ ts/store/reducers/notifications/index.ts
@@ -3,7 +3,7 @@
  */
 
 import { combineReducers } from "redux";
-import { Action } from "../../../actions/types";
+import { Action } from "../../actions/types";
 import installationReducer, { InstallationState } from "./installation";
 
 export type NotificationsState = {
