+++ ts/store/reducers/session.ts
@@ -2,11 +2,11 @@
  * A reducer for the Session.
  */
 
-import { Action } from "../../actions/types";
 import { IdentityProvider } from "../../models/IdentityProvider";
 import { GlobalState } from "../../reducers/types";
 import { SessionToken } from "../../types/SessionToken";
 import { IDP_SELECTED, LOGIN_SUCCESS } from "../actions/constants";
+import { Action } from "../actions/types";
 
 export type UnauthenticatedWithoutIdpSessionState = Readonly<{
   isAuthenticated: false;
